@@ -1,4 +1,5 @@
 ﻿// <auto-generated />
+using System;
 using Application.Data;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
@@ -14,7 +15,7 @@ namespace Application.Data.Migrations
         {
 #pragma warning disable 612, 618
             modelBuilder
-                .HasAnnotation("ProductVersion", "2.1.11-servicing-32099")
+                .HasAnnotation("ProductVersion", "3.0.0")
                 .HasAnnotation("Relational:MaxIdentifierLength", 128)
                 .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
@@ -22,29 +23,40 @@ namespace Application.Data.Migrations
                 {
                     b.Property<int>("AddressId")
                         .ValueGeneratedOnAdd()
+                        .HasColumnType("int")
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
-                    b.Property<int>("AddressType");
+                    b.Property<int>("AddressType")
+                        .HasColumnType("int");
 
-                    b.Property<string>("City");
+                    b.Property<string>("City")
+                        .HasColumnType("nvarchar(max)");
 
                     b.Property<string>("Country")
                         .IsRequired()
+                        .HasColumnType("nvarchar(255)")
                         .HasMaxLength(255);
 
-                    b.Property<int>("EntityState");
+                    b.Property<int>("EntityState")
+                        .HasColumnType("int");
 
-                    b.Property<bool>("HasChanges");
+                    b.Property<bool>("HasChanges")
+                        .HasColumnType("bit");
 
-                    b.Property<bool>("IsNew");
+                    b.Property<bool>("IsNew")
+                        .HasColumnType("bit");
 
-                    b.Property<string>("PostalCode");
+                    b.Property<string>("PostalCode")
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<string>("State");
+                    b.Property<string>("State")
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<string>("StreetLine1");
+                    b.Property<string>("StreetLine1")
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<string>("StreetLine2");
+                    b.Property<string>("StreetLine2")
+                        .HasColumnType("nvarchar(max)");
 
                     b.HasKey("AddressId");
 
@@ -55,20 +67,27 @@ namespace Application.Data.Migrations
                 {
                     b.Property<int>("ID")
                         .ValueGeneratedOnAdd()
+                        .HasColumnType("int")
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
-                    b.Property<int>("AddressId");
+                    b.Property<int>("AddressId")
+                        .HasColumnType("int");
 
-                    b.Property<int>("Cuisine");
+                    b.Property<int>("Cuisine")
+                        .HasColumnType("int");
 
-                    b.Property<int>("EntityState");
+                    b.Property<int>("EntityState")
+                        .HasColumnType("int");
 
-                    b.Property<bool>("HasChanges");
+                    b.Property<bool>("HasChanges")
+                        .HasColumnType("bit");
 
-                    b.Property<bool>("IsNew");
+                    b.Property<bool>("IsNew")
+                        .HasColumnType("bit");
 
                     b.Property<string>("Name")
                         .IsRequired()
+                        .HasColumnType("nvarchar(80)")
                         .HasMaxLength(80);
 
                     b.HasKey("ID");
@@ -78,12 +97,202 @@ namespace Application.Data.Migrations
                     b.ToTable("Restaurants");
                 });
 
+            modelBuilder.Entity("Application.Data.Category", b =>
+                {
+                    b.Property<int>("CategoryId")
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("int")
+                        .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
+
+                    b.Property<string>("CategoryName")
+                        .HasColumnType("nvarchar(max)");
+
+                    b.Property<string>("Description")
+                        .HasColumnType("nvarchar(max)");
+
+                    b.HasKey("CategoryId");
+
+                    b.ToTable("Categories");
+
+                    b.HasData(
+                        new
+                        {
+                            CategoryId = 1,
+                            CategoryName = "Fruit pies",
+                            Description = "All-Fruit Pies"
+                        },
+                        new
+                        {
+                            CategoryId = 2,
+                            CategoryName = "Cheese cakes",
+                            Description = "Cheesy all the way"
+                        },
+                        new
+                        {
+                            CategoryId = 3,
+                            CategoryName = "Seasonal Pies",
+                            Description = "Get in the mood for a seasonal pie"
+                        });
+                });
+
+            modelBuilder.Entity("Application.Data.Pie", b =>
+                {
+                    b.Property<int>("PieId")
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("int")
+                        .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
+
+                    b.Property<string>("AllergyInformation")
+                        .HasColumnType("nvarchar(max)");
+
+                    b.Property<int>("CategoryId")
+                        .HasColumnType("int");
+
+                    b.Property<decimal?>("CurrentPrice")
+                        .HasColumnType("decimal(18,2)");
+
+                    b.Property<int>("EntityState")
+                        .HasColumnType("int");
+
+                    b.Property<bool>("HasChanges")
+                        .HasColumnType("bit");
+
+                    b.Property<string>("ImageThumbnailUrl")
+                        .HasColumnType("nvarchar(max)");
+
+                    b.Property<string>("ImageUrl")
+                        .HasColumnType("nvarchar(max)");
+
+                    b.Property<bool>("InStock")
+                        .HasColumnType("bit");
+
+                    b.Property<bool>("IsNew")
+                        .HasColumnType("bit");
+
+                    b.Property<bool>("IsPieOfTheWeek")
+                        .HasColumnType("bit");
+
+                    b.Property<float>("ListPrice")
+                        .HasColumnType("real");
+
+                    b.Property<string>("LongDescription")
+                        .HasColumnType("nvarchar(max)");
+
+                    b.Property<string>("Name")
+                        .HasColumnType("nvarchar(max)");
+
+                    b.Property<string>("Notes")
+                        .HasColumnType("nvarchar(max)");
+
+                    b.Property<decimal>("Price")
+                        .HasColumnType("decimal(18,2)");
+
+                    b.Property<string>("ProductDescription")
+                        .HasColumnType("nvarchar(max)");
+
+                    b.Property<int>("ProductId")
+                        .HasColumnType("int");
+
+                    b.Property<string>("ShortDescription")
+                        .HasColumnType("nvarchar(max)");
+
+                    b.HasKey("PieId");
+
+                    b.HasIndex("CategoryId");
+
+                    b.ToTable("Pies");
+
+                    b.HasData(
+                        new
+                        {
+                            PieId = 1,
+                            CategoryId = 1,
+                            EntityState = 0,
+                            HasChanges = false,
+                            ImageThumbnailUrl = "https://gillcleerenpluralsight.blob.core.windows.net/files/strawberrypiesmall.jpg",
+                            ImageUrl = "https://gillcleerenpluralsight.blob.core.windows.net/files/strawberrypie.jpg",
+                            InStock = true,
+                            IsNew = false,
+                            IsPieOfTheWeek = false,
+                            ListPrice = 0f,
+                            LongDescription = "Icing carrot cake jelly-o cheesecake. Sweet roll marzipan marshmallow toffee brownie brownie candy tootsie roll. Chocolate cake gingerbread tootsie roll oat cake pie chocolate bar cookie dragée brownie. Lollipop cotton candy cake bear claw oat cake. Dragée candy canes dessert tart. Marzipan dragée gummies lollipop jujubes chocolate bar candy canes. Icing gingerbread chupa chups cotton candy cookie sweet icing bonbon gummies. Gummies lollipop brownie biscuit danish chocolate cake. Danish powder cookie macaroon chocolate donut tart. Carrot cake dragée croissant lemon drops liquorice lemon drops cookie lollipop toffee. Carrot cake carrot cake liquorice sugar plum topping bonbon pie muffin jujubes. Jelly pastry wafer tart caramels bear claw. Tiramisu tart pie cake danish lemon drops. Brownie cupcake dragée gummies.",
+                            Name = "Strawberry Pie",
+                            Price = 15.95m,
+                            ProductId = 0,
+                            ShortDescription = "Lorem Ipsum"
+                        },
+                        new
+                        {
+                            PieId = 2,
+                            CategoryId = 2,
+                            EntityState = 0,
+                            HasChanges = false,
+                            ImageThumbnailUrl = "https://gillcleerenpluralsight.blob.core.windows.net/files/cheesecakesmall.jpg",
+                            ImageUrl = "https://gillcleerenpluralsight.blob.core.windows.net/files/cheesecake.jpg",
+                            InStock = true,
+                            IsNew = false,
+                            IsPieOfTheWeek = false,
+                            ListPrice = 0f,
+                            LongDescription = "Icing carrot cake jelly-o cheesecake. Sweet roll marzipan marshmallow toffee brownie brownie candy tootsie roll. Chocolate cake gingerbread tootsie roll oat cake pie chocolate bar cookie dragée brownie. Lollipop cotton candy cake bear claw oat cake. Dragée candy canes dessert tart. Marzipan dragée gummies lollipop jujubes chocolate bar candy canes. Icing gingerbread chupa chups cotton candy cookie sweet icing bonbon gummies. Gummies lollipop brownie biscuit danish chocolate cake. Danish powder cookie macaroon chocolate donut tart. Carrot cake dragée croissant lemon drops liquorice lemon drops cookie lollipop toffee. Carrot cake carrot cake liquorice sugar plum topping bonbon pie muffin jujubes. Jelly pastry wafer tart caramels bear claw. Tiramisu tart pie cake danish lemon drops. Brownie cupcake dragée gummies.",
+                            Name = "Cheese cake",
+                            Price = 18.95m,
+                            ProductId = 0,
+                            ShortDescription = "Lorem Ipsum"
+                        },
+                        new
+                        {
+                            PieId = 3,
+                            CategoryId = 1,
+                            EntityState = 0,
+                            HasChanges = false,
+                            ImageThumbnailUrl = "https://gillcleerenpluralsight.blob.core.windows.net/files/rhubarbpiesmall.jpg",
+                            ImageUrl = "https://gillcleerenpluralsight.blob.core.windows.net/files/rhubarbpie.jpg",
+                            InStock = true,
+                            IsNew = false,
+                            IsPieOfTheWeek = true,
+                            ListPrice = 0f,
+                            LongDescription = "Icing carrot cake jelly-o cheesecake. Sweet roll marzipan marshmallow toffee brownie brownie candy tootsie roll. Chocolate cake gingerbread tootsie roll oat cake pie chocolate bar cookie dragée brownie. Lollipop cotton candy cake bear claw oat cake. Dragée candy canes dessert tart. Marzipan dragée gummies lollipop jujubes chocolate bar candy canes. Icing gingerbread chupa chups cotton candy cookie sweet icing bonbon gummies. Gummies lollipop brownie biscuit danish chocolate cake. Danish powder cookie macaroon chocolate donut tart. Carrot cake dragée croissant lemon drops liquorice lemon drops cookie lollipop toffee. Carrot cake carrot cake liquorice sugar plum topping bonbon pie muffin jujubes. Jelly pastry wafer tart caramels bear claw. Tiramisu tart pie cake danish lemon drops. Brownie cupcake dragée gummies.",
+                            Name = "Rhubarb Pie",
+                            Price = 15.95m,
+                            ProductId = 0,
+                            ShortDescription = "Lorem Ipsum"
+                        },
+                        new
+                        {
+                            PieId = 4,
+                            CategoryId = 3,
+                            EntityState = 0,
+                            HasChanges = false,
+                            ImageThumbnailUrl = "https://gillcleerenpluralsight.blob.core.windows.net/files/pumpkinpiesmall.jpg",
+                            ImageUrl = "https://gillcleerenpluralsight.blob.core.windows.net/files/pumpkinpie.jpg",
+                            InStock = true,
+                            IsNew = false,
+                            IsPieOfTheWeek = true,
+                            ListPrice = 0f,
+                            LongDescription = "Icing carrot cake jelly-o cheesecake. Sweet roll marzipan marshmallow toffee brownie brownie candy tootsie roll. Chocolate cake gingerbread tootsie roll oat cake pie chocolate bar cookie dragée brownie. Lollipop cotton candy cake bear claw oat cake. Dragée candy canes dessert tart. Marzipan dragée gummies lollipop jujubes chocolate bar candy canes. Icing gingerbread chupa chups cotton candy cookie sweet icing bonbon gummies. Gummies lollipop brownie biscuit danish chocolate cake. Danish powder cookie macaroon chocolate donut tart. Carrot cake dragée croissant lemon drops liquorice lemon drops cookie lollipop toffee. Carrot cake carrot cake liquorice sugar plum topping bonbon pie muffin jujubes. Jelly pastry wafer tart caramels bear claw. Tiramisu tart pie cake danish lemon drops. Brownie cupcake dragée gummies.",
+                            Name = "Pumpkin Pie",
+                            Price = 12.95m,
+                            ProductId = 0,
+                            ShortDescription = "Lorem Ipsum"
+                        });
+                });
+
             modelBuilder.Entity("ACM.BL.Restaurant", b =>
                 {
                     b.HasOne("ACM.BL.Address", "Address")
                         .WithMany()
                         .HasForeignKey("AddressId")
-                        .OnDelete(DeleteBehavior.Cascade);
+                        .OnDelete(DeleteBehavior.Cascade)
+                        .IsRequired();
+                });
+
+            modelBuilder.Entity("Application.Data.Pie", b =>
+                {
+                    b.HasOne("Application.Data.Category", "Category")
+                        .WithMany("Pies")
+                        .HasForeignKey("CategoryId")
+                        .OnDelete(DeleteBehavior.Cascade)
+                        .IsRequired();
                 });
 #pragma warning restore 612, 618
         }
