@@ -22,11 +22,15 @@ namespace Application.Data
         public int ProductId { get; private set; }
         public string ShortDescription { get; set; }
         public string LongDescription { get; set; }
+        public decimal Price { get; set; }
         public string ImageUrl { get; set; }
         public string ImageThumbnailUrl { get; set; }
-        public Category Category { get; set; }
-        public decimal Price { get; set; }
+
         public bool InStock { get; set; }
+        public int CategoryId { get; set; }
+        public Category Category { get; set; }
+        public string Notes { get; set; }
+
 
         private string _name;
         public string Name
