@@ -1,9 +1,6 @@
-﻿using ACM.BL;
-using Application.Data;
-using System;
+﻿using Application.Data;
 using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+
 
 namespace BusinessSample.ViewModels
 {
